1658 - update linked list tail pointer correctly
Given the way this value was used this plug-in should have behaved
correctly but the variable name was not reflecting its actual use.

I also freed a variable that would leak memory in an error case,
again this is not an issue with the functionality of the plug-in as
the error case will cause an immediate clean shutdown but it was
not strictly correct.

Correct several typos in man pages.