Compile semi-static sample plugins (#16, #2096)
Compile mistral_mysql and mistral_influxdb sample plugins for packaging. Although we refer
sometimes refer to these as statically linked, only some libraries are statically linked, with the
more "standard" libraries still being dynamically linked.

This is very much a work in progresss! At this point we can remotely compile on 10.33.0.101 and
10.33.0.102 and get back plausible looking plugins, but as yet these aren't actually packaged up!

This relates to issue #16 in the mistral_plugins repository and issue #2096 in the breeze-repo
repository.