Makefile shodul continue if the statc link fails (#16)
Modify Makefiles so that if the static link fails we'll carry on processing. This can happen when
running on a machine that doesn't have the necessary (static) libraries installed. The build
machines have all necessary libraries so shouldn't fail, but if they do then the error will be
seen when we try to copy back the compiled plugin. This approach means that people can build
dynamically linked plugins when developing code, and then do a final build using the package
plugins script to get a plugin that can run anywhere.

(mistral_plugins #16, breeze-repo #2096)