Various fixes / minor enhancements to influx plug-in
This change pushes some generic framework updates that make sense
in terms of requirements from customers i.e. storing hostname
with and without the domain, only appending a new line to an error
if one is not already present.

It also has some bug fixes, MPI rank may be null but the log parsing
did not allow that, curl queries were using very broken options yet
we working by pure luck.

Test framework changes, I'm fairly sure these are not needed here
but as a generic fix it would be nice to have these back in master.

I have also included some debug output similar to what is in place
for the RTM plug-in as I added it for figuring out a problem I was
seeing and it doesn't hurt anything (as far as I can tell!)

Finally I added a line to the expected test output that is present
in newer InfluxDB versions.