Mistral logs now include sequence number (#2234)
With the ability to save stack backtraces in a plain text log file,
Mistral log messages now include a sequence number which can be used to
identify the backtrace associated with any given log message.

The common code has been extended to parse and save the sequence number,
but the value is not yet used by any of the plugins.

The input data for each test has had a seqeunce number appended to each
line, to mimic what Mistral would have produced.

The mistral_rtm input data hasn't been modified, on the assumption that
changes need to be coordinated with IBM.