Improvement for reading data from Mistral
This is related with an issue that was reported for Mistral (which is Plugin not ready to receive data #3852).

In the previous implementation reading from stdin was unbuffered. This means that the plug-in was reading the data sent by Mistral one character at a time.

Note that this new implementation does not fix by itself the issue #3852 but improves things significantly.